backend/requirements.txt
spotipy
fastapi
uvicorn[standard]
python-dotenv
google-adk
google-cloud-aiplatform
starlette
python-multipart
itsdangerous